+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
-## Unreleased
+## 2.5.0 [2021-02-20]
+
+### Added
+
+- transport configurations support
 
 ### Fixed